requirements.txt
# flask
# tensorflow
# opencv-python
# numpy
# Pillow


asttokens==3.0.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
comm==0.2.2
debugpy==1.8.9
decorator==5.1.1
executing==2.1.0
filelock==3.18.0
Flask==3.1.0
fsspec==2025.3.0
ipykernel==6.29.5
ipython==8.30.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==308
pyzmq==26.2.0
setuptools==78.1.0
six==1.17.0
stack-data==0.6.3
sympy==1.13.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.0
tzdata==2024.2
wcwidth==0.2.13
Werkzeug==3.1.3
